Make `RoutingWrapper` work with async
